+++ source/lib/output/csv.hpp
@@ -99,17 +99,18 @@ struct csv_encoder
     }
 };
 
-using api_csv_encoder                  = csv_encoder<7>;
-using agent_info_csv_encoder           = csv_encoder<53>;
-using kernel_trace_csv_encoder         = csv_encoder<18>;
-using counter_collection_csv_encoder   = csv_encoder<18>;
-using memory_copy_csv_encoder          = csv_encoder<7>;
-using memory_allocation_csv_encoder    = csv_encoder<8>;
-using marker_csv_encoder               = csv_encoder<7>;
-using list_basic_metrics_csv_encoder   = csv_encoder<5>;
-using list_derived_metrics_csv_encoder = csv_encoder<5>;
-using scratch_memory_encoder           = csv_encoder<8>;
-using stats_csv_encoder                = csv_encoder<8>;
+using api_csv_encoder                   = csv_encoder<7>;
+using agent_info_csv_encoder            = csv_encoder<53>;
+using kernel_trace_csv_encoder          = csv_encoder<18>;
+using counter_collection_csv_encoder    = csv_encoder<18>;
+using memory_copy_csv_encoder           = csv_encoder<7>;
+using memory_allocation_csv_encoder     = csv_encoder<8>;
+using marker_csv_encoder                = csv_encoder<7>;
+using list_basic_metrics_csv_encoder    = csv_encoder<5>;
+using list_derived_metrics_csv_encoder  = csv_encoder<5>;
+using scratch_memory_encoder            = csv_encoder<8>;
+using stats_csv_encoder                 = csv_encoder<8>;
+using pc_sampling_host_trap_csv_encoder = csv_encoder<6>;
 }  // namespace csv
 }  // namespace tool
 }  // namespace rocprofiler