+++ source/lib/output/generateStats.hpp
@@ -75,6 +75,10 @@ generate_stats(const output_config& cfg,
                const metadata&      tool_metadata,
                const generator<rocprofiler_buffer_tracing_memory_allocation_record_t>& data);
 
+stats_entry_t
+generate_stats(const output_config&                                              cfg,
+               const metadata&                                                   tool_metadata,
+               const generator<rocprofiler_tool_pc_sampling_host_trap_record_t>& data);
 void
 generate_stats(const output_config&      cfg,
                const metadata&           tool_metadata,